collector/requirements.txt
websockets==12.0
aiokafka==0.11.0
pydantic==2.7.0
pyyaml==6.0.1
lz4==4.3.3
